+++ app/controllers/api/account.php
@@ -81,9 +81,7 @@ App::post('/v1/account')
         $limit = $project->getAttribute('auths', [])['limit'] ?? 0;
 
         if ($limit !== 0) {
-            $total = $dbForProject->count('users', [
-                new Query('deleted', Query::TYPE_EQUAL, [false]),
-            ], APP_LIMIT_USERS);
+            $total = $dbForProject->count('users', max: APP_LIMIT_USERS);
 
             if ($total >= $limit) {
                 throw new Exception('Project registration is restricted. Contact your administrator for more information.', 501, Exception::USER_COUNT_EXCEEDED);
@@ -108,8 +106,7 @@ App::post('/v1/account')
                 'sessions' => null,
                 'tokens' => null,
                 'memberships' => null,
-                'search' => implode(' ', [$userId, $email, $name]),
-                'deleted' => false
+                'search' => implode(' ', [$userId, $email, $name])
             ])));
         } catch (Duplicate $th) {
             throw new Exception('Account already exists', 409, Exception::USER_ALREADY_EXISTS);
@@ -170,7 +167,6 @@ App::post('/v1/account/sessions')
         $protocol = $request->getProtocol();
 
         $profile = $dbForProject->findOne('users', [
-            new Query('deleted', Query::TYPE_EQUAL, [false]),
             new Query('email', Query::TYPE_EQUAL, [$email])]
         );
 
@@ -482,7 +478,6 @@ App::get('/v1/account/sessions/oauth2/:provider/redirect')
             $isVerified = $oauth2->isEmailVerified($accessToken);
 
             $user = $dbForProject->findOne('users', [
-                new Query('deleted', Query::TYPE_EQUAL, [false]),
                 new Query('email', Query::TYPE_EQUAL, [$email])]
             );
 
@@ -490,7 +485,7 @@ App::get('/v1/account/sessions/oauth2/:provider/redirect')
                 $limit = $project->getAttribute('auths', [])['limit'] ?? 0;
 
                 if ($limit !== 0) {
-                    $total = $dbForProject->count('users', [new Query('deleted', Query::TYPE_EQUAL, [false])], APP_LIMIT_USERS);
+                    $total = $dbForProject->count('users', max: APP_LIMIT_USERS);
 
                     if ($total >= $limit) {
                         throw new Exception('Project registration is restricted. Contact your administrator for more information.', 501, Exception::USER_COUNT_EXCEEDED);
@@ -515,8 +510,7 @@ App::get('/v1/account/sessions/oauth2/:provider/redirect')
                         'sessions' => null,
                         'tokens' => null,
                         'memberships' => null,
-                        'search' => implode(' ', [$userId, $email, $name]),
-                        'deleted' => false
+                        'search' => implode(' ', [$userId, $email, $name])
                     ])));
                 } catch (Duplicate $th) {
                     throw new Exception('Account already exists', 409, Exception::USER_ALREADY_EXISTS);
@@ -664,9 +658,7 @@ App::post('/v1/account/sessions/magic-url')
             $limit = $project->getAttribute('auths', [])['limit'] ?? 0;
 
             if ($limit !== 0) {
-                $total = $dbForProject->count('users', [
-                    new Query('deleted', Query::TYPE_EQUAL, [false]),
-                ], APP_LIMIT_USERS);
+                $total = $dbForProject->count('users', max: APP_LIMIT_USERS);
 
                 if ($total >= $limit) {
                     throw new Exception('Project registration is restricted. Contact your administrator for more information.', 501, Exception::USER_COUNT_EXCEEDED);
@@ -690,8 +682,7 @@ App::post('/v1/account/sessions/magic-url')
                 'sessions' => null,
                 'tokens' => null,
                 'memberships' => null,
-                'search' => implode(' ', [$userId, $email]),
-                'deleted' => false
+                'search' => implode(' ', [$userId, $email])
             ])));
         }
 
@@ -791,7 +782,7 @@ App::put('/v1/account/sessions/magic-url')
 
         $user = Authorization::skip(fn() => $dbForProject->getDocument('users', $userId));
 
-        if ($user->isEmpty() || $user->getAttribute('deleted')) {
+        if ($user->isEmpty()) {
             throw new Exception('User not found', 404, Exception::USER_NOT_FOUND);
         }
 
@@ -926,9 +917,7 @@ App::post('/v1/account/sessions/anonymous')
         $limit = $project->getAttribute('auths', [])['limit'] ?? 0;
 
         if ($limit !== 0) {
-            $total = $dbForProject->count('users', [
-                new Query('deleted', Query::TYPE_EQUAL, [false]),
-            ], APP_LIMIT_USERS);
+            $total = $dbForProject->count('users', max: APP_LIMIT_USERS);
 
             if ($total >= $limit) {
                 throw new Exception('Project registration is restricted. Contact your administrator for more information.', 501, Exception::USER_COUNT_EXCEEDED);
@@ -952,8 +941,7 @@ App::post('/v1/account/sessions/anonymous')
             'sessions' => null,
             'tokens' => null,
             'memberships' => null,
-            'search' => $userId,
-            'deleted' => false
+            'search' => $userId
         ])));
 
         // Create session token
@@ -1470,17 +1458,18 @@ App::patch('/v1/account/prefs')
         $response->dynamic($user, Response::MODEL_USER);
     });
 
-App::delete('/v1/account')
-    ->desc('Delete Account')
+App::patch('/v1/account/status')
+    ->desc('Update Account Status')
     ->groups(['api', 'account'])
-    ->label('event', 'users.[userId].delete')
+    ->label('event', 'users.[userId].update.status')
     ->label('scope', 'account')
     ->label('sdk.auth', [APP_AUTH_TYPE_SESSION, APP_AUTH_TYPE_JWT])
     ->label('sdk.namespace', 'account')
-    ->label('sdk.method', 'delete')
-    ->label('sdk.description', '/docs/references/account/delete.md')
-    ->label('sdk.response.code', Response::STATUS_CODE_NOCONTENT)
-    ->label('sdk.response.model', Response::MODEL_NONE)
+    ->label('sdk.method', 'updateStatus')
+    ->label('sdk.description', '/docs/references/account/update-status.md')
+    ->label('sdk.response.code', Response::STATUS_CODE_OK)
+    ->label('sdk.response.type', Response::CONTENT_TYPE_JSON)
+    ->label('sdk.response.model', Response::MODEL_USER)
     ->inject('request')
     ->inject('response')
     ->inject('user')
@@ -1497,28 +1486,15 @@ App::delete('/v1/account')
         /** @var Appwrite\Event\Event $events */
         /** @var Appwrite\Stats\Stats $usage */
 
-        $protocol = $request->getProtocol();
         $user = $dbForProject->updateDocument('users', $user->getId(), $user->setAttribute('status', false));
-
-        // TODO Seems to be related to users.php/App::delete('/v1/users/:userId'). Can we share code between these two? Do todos below apply to users.php?
-
-        // TODO delete all tokens or only current session?
-        // TODO delete all user data according to GDPR. Make sure everything is backed up and backups are deleted later
-        /**
-        * Data to delete
-        * * Tokens
-        * * Memberships
-        */
 
         $audits
             ->setResource('user/' . $user->getId())
-            ->setPayload($response->output($user, Response::MODEL_USER))
-        ;
+            ->setPayload($response->output($user, Response::MODEL_USER));
 
         $events
             ->setParam('userId', $user->getId())
-            ->setPayload($response->output($user, Response::MODEL_USER))
-        ;
+            ->setPayload($response->output($user, Response::MODEL_USER));
 
         if (!Config::getParam('domainVerification')) {
             $response->addHeader('X-Fallback-Cookies', \json_encode([]));
@@ -1526,11 +1502,7 @@ App::delete('/v1/account')
 
         $usage->setParam('users.delete', 1);
 
-        $response
-            ->addCookie(Auth::$cookieName . '_legacy', '', \time() - 3600, '/', Config::getParam('cookieDomain'), ('https' == $protocol), true, null)
-            ->addCookie(Auth::$cookieName, '', \time() - 3600, '/', Config::getParam('cookieDomain'), ('https' == $protocol), true, Config::getParam('cookieSamesite'))
-            ->noContent()
-        ;
+        $response->dynamic($user, Response::MODEL_USER);
     });
 
 App::delete('/v1/account/sessions/:sessionId')
@@ -1840,7 +1812,6 @@ App::post('/v1/account/recovery')
         $email = \strtolower($email);
 
         $profile = $dbForProject->findOne('users', [
-            new Query('deleted', Query::TYPE_EQUAL, [false]), 
             new Query('email', Query::TYPE_EQUAL, [$email])
         ]);
 
@@ -1943,7 +1914,7 @@ App::put('/v1/account/recovery')
 
         $profile = $dbForProject->getDocument('users', $userId);
 
-        if ($profile->isEmpty() || $profile->getAttribute('deleted')) {
+        if ($profile->isEmpty()) {
             throw new Exception('User not found', 404, Exception::USER_NOT_FOUND);
         }